Move Codable to Foundation Extension
